fix(binding): skip unix mode audit on windows
